Refactor HTMLBars compiler into lib/ folder
